+++ internal/openvpn/connection/client_test.go
@@ -1,10 +1,10 @@
-package openvpn_test
+package connection_test
 
 import (
 	"strings"
 	"testing"
 
-	"github.com/jkroepke/openvpn-auth-oauth2/internal/openvpn"
+	"github.com/jkroepke/openvpn-auth-oauth2/internal/openvpn/connection"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -14,13 +14,13 @@ func TestNewClientConnection(t *testing.T) {
 	for _, tt := range []struct {
 		name             string
 		lines            []string
-		clientConnection openvpn.ClientConnection
+		clientConnection connection.Client
 		err              string
 	}{
 		{
 			"client CONNECT",
 			[]string{">CLIENT:CONNECT,0,1", ">CLIENT:ENV,name1=val1", ">CLIENT:ENV,name2=", ">CLIENT:ENV,END"},
-			openvpn.ClientConnection{Cid: 0, Kid: 1, Reason: "CONNECT", Env: map[string]string{
+			connection.Client{Cid: 0, Kid: 1, Reason: "CONNECT", Env: map[string]string{
 				"name1": "val1", "name2": "",
 			}},
 			"",
@@ -28,7 +28,7 @@ func TestNewClientConnection(t *testing.T) {
 		{
 			"client CONNECT invalid cid",
 			[]string{">CLIENT:CONNECT,k,2", ">CLIENT:ENV,name1=val1", ">CLIENT:ENV,name2=", ">CLIENT:ENV,END"},
-			openvpn.ClientConnection{Cid: 1, Kid: 2, Reason: "CONNECT", Env: map[string]string{
+			connection.Client{Cid: 1, Kid: 2, Reason: "CONNECT", Env: map[string]string{
 				"name1": "val1", "name2": "",
 			}},
 			"unable to parse cid: strconv.ParseUint: parsing \"k\": invalid syntax",
@@ -36,7 +36,7 @@ func TestNewClientConnection(t *testing.T) {
 		{
 			"client CONNECT invalid kid",
 			[]string{">CLIENT:CONNECT,1,k", ">CLIENT:ENV,name1=val1", ">CLIENT:ENV,name2=", ">CLIENT:ENV,END"},
-			openvpn.ClientConnection{Cid: 1, Kid: 2, Reason: "CONNECT", Env: map[string]string{
+			connection.Client{Cid: 1, Kid: 2, Reason: "CONNECT", Env: map[string]string{
 				"name1": "val1", "name2": "",
 			}},
 			"unable to parse kid: strconv.ParseUint: parsing \"k\": invalid syntax",
@@ -44,7 +44,7 @@ func TestNewClientConnection(t *testing.T) {
 		{
 			"client REAUTH",
 			[]string{">CLIENT:REAUTH,1,2", ">CLIENT:ENV,name1=val1", ">CLIENT:ENV,name2=", ">CLIENT:ENV,END"},
-			openvpn.ClientConnection{Cid: 1, Kid: 2, Reason: "REAUTH", Env: map[string]string{
+			connection.Client{Cid: 1, Kid: 2, Reason: "REAUTH", Env: map[string]string{
 				"name1": "val1", "name2": "",
 			}},
 			"",
@@ -52,7 +52,7 @@ func TestNewClientConnection(t *testing.T) {
 		{
 			"client ESTABLISHED",
 			[]string{">CLIENT:ESTABLISHED,1", ">CLIENT:ENV,name1=val1", ">CLIENT:ENV,name2=", ">CLIENT:ENV,END"},
-			openvpn.ClientConnection{
+			connection.Client{
 				Cid: 1, Kid: 0, Reason: "ESTABLISHED", Env: map[string]string{
 					"name1": "val1", "name2": "",
 				},
@@ -62,7 +62,7 @@ func TestNewClientConnection(t *testing.T) {
 		{
 			"client DISCONNECT",
 			[]string{">CLIENT:DISCONNECT,1", ">CLIENT:ENV,name1=val1", ">CLIENT:ENV,name2=", ">CLIENT:ENV,END"},
-			openvpn.ClientConnection{
+			connection.Client{
 				Cid: 1, Kid: 0, Reason: "DISCONNECT", Env: map[string]string{
 					"name1": "val1", "name2": "",
 				},
@@ -72,7 +72,7 @@ func TestNewClientConnection(t *testing.T) {
 		{
 			"client CR_RESPONSE",
 			[]string{">CLIENT:CR_RESPONSE,1,2,YmFzZTY0", ">CLIENT:ENV,name1=val1", ">CLIENT:ENV,END"},
-			openvpn.ClientConnection{
+			connection.Client{
 				Cid: 1, Kid: 2, Reason: "CR_RESPONSE", Env: map[string]string{
 					"name1": "val1",
 				},
@@ -82,13 +82,13 @@ func TestNewClientConnection(t *testing.T) {
 		{
 			"client invalid reason",
 			[]string{">CLIENT:unknown", ">CLIENT:ENV,name1=val1", ">CLIENT:ENV,name2", ">CLIENT:ENV,END"},
-			openvpn.ClientConnection{},
+			connection.Client{},
 			"unable to parse client reason from message: >CLIENT:unknown\n>CLIENT:ENV,name1=val1\n>CLIENT:ENV,name2\n>CLIENT:ENV,END",
 		},
 		{
 			"client invalid reason",
 			[]string{">CLIENT:CONNECT", ">CLIENT:ENV,name1=val1", ">CLIENT:ENV,name2", ">CLIENT:ENV,END"},
-			openvpn.ClientConnection{},
+			connection.Client{},
 			"unable to parse line '>CLIENT:CONNECT': message invalid",
 		},
 	} {
@@ -99,7 +99,7 @@ func TestNewClientConnection(t *testing.T) {
 
 			message := strings.Join(tt.lines, "\n")
 
-			clientConnection, err := openvpn.NewClientConnection(message)
+			clientConnection, err := connection.NewClient(message)
 			if tt.err == "" {
 				assert.NoError(t, err)
 				assert.Equal(t, tt.clientConnection, clientConnection)
