+++ internal/openvpn/connection/errors.go
@@ -1,0 +1,10 @@
+package connection
+
+import (
+	"errors"
+)
+
+var (
+	ErrEmptyClientReasons = errors.New("empty client reason")
+	ErrInvalidMessage     = errors.New("message invalid")
+)
